% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-package.R
\docType{package}
\name{pandoc-package}
\alias{pandoc}
\alias{pandoc-package}
\title{pandoc: Manage and Run Universal Converter 'Pandoc' from 'R'}
\description{
Provides a set of tools to install, manage and run several 'Pandoc' versions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cderv/pandoc}
  \item \url{https://cderv.github.io/pandoc/}
  \item Report bugs at \url{https://github.com/cderv/pandoc/issues}
}

}
\author{
\strong{Maintainer}: Christophe Dervieux \email{cderv@rstudio.com} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
