% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_citation}
\alias{pandoc_browse_citation}
\title{Open Pandoc's documentation about citation processing}
\usage{
pandoc_browse_citation()
}
\value{
Open the webpage at the place regarding citation processing in Pandoc.
}
\description{
Open Pandoc's documentation about citation processing
}
\examples{
\dontshow{
withr::local_options(list(rlang_interactive = FALSE))
}
pandoc_browse_citation()
}
