% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_exit_code}
\alias{pandoc_browse_exit_code}
\title{Open Pandoc's documentation about exit codes}
\usage{
pandoc_browse_exit_code()
}
\value{
Open the webpage at the place regarding exit code thrown by Pandoc.
}
\description{
Open Pandoc's documentation about exit codes
}
\examples{
\dontshow{
withr::local_options(list(rlang_interactive = FALSE))
}
pandoc_browse_exit_code()
}
