% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_extension}
\alias{pandoc_browse_extension}
\title{Open Pandoc's documentation about an extension}
\usage{
pandoc_browse_extension(extension)
}
\arguments{
\item{extension}{One of the supported extension. See
\code{\link[=pandoc_list_extensions]{pandoc_list_extensions()}}. As the Pandoc MANUAL only concerns the last
released Pandoc's version, if the URL is incorrect this could mean the
extensions has changed.}
}
\value{
Open the webpage at the place regarding the required extension.
}
\description{
Open Pandoc's documentation about an extension
}
