% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_manual}
\alias{pandoc_browse_manual}
\title{Open Pandoc's MANUAL}
\usage{
pandoc_browse_manual(id = NULL)
}
\arguments{
\item{id}{One of the id available in the HTML page (usually for anchor link).}
}
\value{
Open the Pandoc's MANUAL
}
\description{
Open Pandoc's MANUAL
}
\examples{
# open MANUAL home page
pandoc_browse_manual()
# open MANUAL at math part
pandoc_browse_manual("math")
}
\references{
\url{https://pandoc.org/MANUAL.html}
}
