% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_option}
\alias{pandoc_browse_option}
\title{Open Pandoc's documentation about a command line option}
\usage{
pandoc_browse_option(option = NULL)
}
\arguments{
\item{option}{One of the supported \strong{long form} command line option. As the
Pandoc MANUAL only concerns the last released Pandoc's version, if the URL
is incorrect this could mean the option has changed.}
}
\value{
Open the webpage at the place regarding the required option
}
\description{
Open Pandoc's documentation about a command line option
}
