% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-template.R
\name{pandoc_get_data_file}
\alias{pandoc_get_data_file}
\title{Export Pandoc internal data file}
\usage{
pandoc_get_data_file(file, output = file, bin = pandoc_bin())
}
\arguments{
\item{file}{One of data file name included in Pandoc (e.g \code{reference.pptx}, \code{styles.html})}

\item{output}{Path where to export the file. Default to working directory
with the same file name.}

\item{bin}{Path to a pandoc binary. Default to the active version.}
}
\description{
This correspond to the \href{https://pandoc.org/MANUAL.html#option--print-default-data-file}{\code{--print-default-data-file} CLI flag} using
also \code{--output} to write a export a data file build in Pandoc.
}
