% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-template.R
\name{pandoc_get_template}
\alias{pandoc_get_template}
\title{Retrieve Pandoc template for a format}
\usage{
pandoc_get_template(format = "markdown", output = NULL, bin = pandoc_bin())
}
\arguments{
\item{format}{One of Pandoc format using a text template. (e.g html, latex, revealjs)}

\item{output}{Path where to save the file. If not provided, the default, template content will be print to the console.}

\item{bin}{Path to a pandoc binary. Default to the active version.}
}
\value{
If \code{output} is not provided, the content of the template will be
printed and return as one string (invisibly). If \code{output} is provided, the file path of
the output (invisibly).
}
\description{
This correspond to the \href{https://pandoc.org/MANUAL.html#option--print-default-template}{\code{--print-default-template} CLI flag}. With
this function, one can easily export default LaTeX template for example.
}
