% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{pandoc_get_version}
\alias{pandoc_get_version}
\title{Get Pandoc version}
\usage{
pandoc_get_version(bin = pandoc_bin())
}
\arguments{
\item{bin}{Path to a pandoc binary. Default to the active version}
}
\description{
This is equivalent to \code{pandoc --version}
}
