% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_home}
\alias{pandoc_home}
\title{Path version Pandoc binaries are installed}
\usage{
pandoc_home(version = NULL)
}
\arguments{
\item{version}{\code{NULL} (the default) will return the install folder for all version. If specified}
}
\description{
Path version Pandoc binaries are installed
}
