% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_home_dir}
\alias{pandoc_home_dir}
\title{Path where Pandoc binaries are installed}
\usage{
pandoc_home_dir(version = "latest")
}
\arguments{
\item{version}{\code{"latest"} (the default) will return the path to directory of
the latest installed Pandoc version. If a version
is specified (e.g \code{"2.11.4"}), it will return the path for a specific
version if installed. If \code{NULL}, the main path to directory where all versions are stored.}
}
\value{
Path of Pandoc binaries root folder if installed.
}
\description{
Path where Pandoc binaries are installed
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
