% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_installed_versions}
\alias{pandoc_installed_versions}
\alias{pandoc_is_installed}
\title{Check Pandoc versions already installed}
\usage{
pandoc_installed_versions()

pandoc_is_installed(version)
}
\arguments{
\item{version}{This can be either:
\itemize{
\item \code{"latest"} for the latest release
\item A version number (e.g \code{"2.11.4"}) for a specific version
\item \code{"nightly"} for the last pandoc development built daily
}}
}
\value{
A character vector of available installed versions or a logical for
\code{pandoc_is_installed()}
}
\description{
\itemize{
\item \code{pandoc_installed_versions()} will list all versions already installed
\item \code{pandoc_is_installed()} allows to check for a specific installed version
}
}
