% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-bin.R
\name{quarto_pandoc}
\alias{quarto_pandoc}
\alias{pandoc_quarto_version}
\alias{pandoc_quarto_bin}
\title{Retrieve path and version of Pandoc shipped with Quarto CLI}
\usage{
pandoc_quarto_version()

pandoc_quarto_bin()
}
\value{
\code{pandoc_quarto_version()} returns the version number for \code{pandoc} binary used by Quarto CLI as a \code{\link[base:numeric_version]{base::numeric_version()}} object.

\code{pandoc_quarto_bin()} returns absolute path to the \code{pandoc} binary used by Quarto CLI.
}
\description{
Quarto CLI ships with a pandoc binary. The path can be found by running
\code{quarto --paths} command. These functions are helpers to easily use this specific version.
}
\examples{
\dontshow{if (!is.null(pandoc::pandoc_quarto_bin())) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
pandoc_quarto_bin()
}
\seealso{
\code{\link[=pandoc_version]{pandoc_version()}}, \code{\link[=pandoc_bin]{pandoc_bin()}}
}
