% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-bin.R
\name{rstudio_pandoc}
\alias{rstudio_pandoc}
\alias{pandoc_rstudio_version}
\alias{pandoc_rstudio_bin}
\title{Retrieve path and version of Pandoc shipped with RStudio}
\usage{
pandoc_rstudio_version()

pandoc_rstudio_bin()
}
\value{
\code{pandoc_rstudio_version()} returns the version number for \code{pandoc} binary used by RStudio IDE as a \code{\link[base:numeric_version]{base::numeric_version()}} object.

\code{pandoc_system_bin()} returns absolute path to the \code{pandoc} binary used by RStudio IDE.
}
\description{
RStudio IDE ships with a pandoc binary. The path is stored in \code{RSTUDIO_PANDOC}
environment variable. Theses function are helper to easily use this specific version.
}
\examples{
\dontshow{if (!is.null(pandoc::pandoc_rstudio_bin())) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
pandoc_rstudio_bin()
}
\seealso{
\code{\link[=pandoc_version]{pandoc_version()}}, \code{\link[=pandoc_bin]{pandoc_bin()}}
}
