% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-bin.R
\name{system_pandoc}
\alias{system_pandoc}
\alias{pandoc_system_version}
\alias{pandoc_system_bin}
\title{Retrieve path and version of Pandoc found on the system PATH}
\usage{
pandoc_system_version()

pandoc_system_bin()
}
\value{
\code{pandoc_system_version()} returns the version number for \code{pandoc} binary found in PATH as a \code{\link[base:numeric_version]{base::numeric_version()}} object.

\code{pandoc_system_bin()} returns absolute path to the \code{pandoc} binary found in PATH.
}
\description{
Pandoc can also be installed on a system and available through the PATH.
Theses function are helper to easily use this specific version.
}
\examples{
\dontshow{if (!is.null(pandoc::pandoc_system_bin())) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
pandoc_system_bin()
}
\seealso{
\code{\link[=pandoc_version]{pandoc_version()}}, \code{\link[=pandoc_bin]{pandoc_bin()}}
}
