% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-available.R
\name{pandoc_bin}
\alias{pandoc_bin}
\alias{pandoc_citeproc_bin}
\title{Get path to the pandoc binary}
\usage{
pandoc_bin(version = "default")

pandoc_citeproc_bin(version = "default")
}
\arguments{
\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}
}
\value{
Absolute path to the pandoc binary of the requested version.

For \code{pandoc_citeproc_bin()}, it returns the path to \code{pandoc-citeproc} binary
if it exists. Since Pandoc 2.11, the citeproc filter has been included into
Pandoc itself and is no more shipped as a binary filter.
}
\description{
Get path to the pandoc binary
}
