% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-bin.R
\name{pandoc_bin}
\alias{pandoc_bin}
\alias{pandoc_bin_browse}
\title{Get path to the pandoc binary}
\usage{
pandoc_bin(version = "default")

pandoc_bin_browse(version = "default")
}
\arguments{
\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}
}
\value{
Absolute path to the pandoc binary of the requested version.
}
\description{
Get path to the pandoc binary
}
\details{
\code{pandoc_bin_browse()} allows to open in OS explorer the folder where
\code{pandoc_bin()} is at, when in interactive mode only.
}
