% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{pandoc_bin}
\alias{pandoc_bin}
\title{Get path to the pandoc binary}
\usage{
pandoc_bin(version = "default")
}
\arguments{
\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version e.g \code{"nightly"}
\item The latest installed version \code{"latest"}
}}
}
\value{
Absolute path to the pandoc binary of the requested version
}
\description{
Get path to the pandoc binary
}
