% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-browse.R
\name{pandoc_browse_release}
\alias{pandoc_browse_release}
\title{Open Pandoc's release page in browser}
\usage{
pandoc_browse_release(version = "latest")
}
\arguments{
\item{version}{One of pandoc release version number (e.g '2.11.2') or \code{"latest"}.}
}
\value{
Open the web page in browser in interactive mode or print the url
}
\description{
Open Pandoc's release page in browser
}
\examples{
\dontshow{if (rlang::is_interactive() && rlang::is_installed("gh") && attr(curlGetHeaders("https://github.com"), "status") == "200") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pandoc_browse_release()
pandoc_browse_release("2.14")
\dontshow{\}) # examplesIf}
}
\references{
\url{https://github.com/jgm/pandoc/releases}
}
