% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-template.R
\name{pandoc_get_data_file}
\alias{pandoc_get_data_file}
\title{Export Pandoc internal data file}
\usage{
pandoc_get_data_file(file, output = file, version = "default")
}
\arguments{
\item{file}{One of data file name included in Pandoc (e.g \code{reference.pptx}, \code{styles.html})}

\item{output}{Path where to export the file. Default to working directory
with the same file name.}

\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}
}
\description{
This correspond to the \href{https://pandoc.org/MANUAL.html#option--print-default-data-file}{\code{--print-default-data-file} CLI flag} using
also \code{--output} to write a export a data file build in Pandoc.
}
