% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_home}
\alias{pandoc_home}
\title{Path where Pandoc binaries are installed}
\usage{
pandoc_home(version = NULL)
}
\arguments{
\item{version}{\code{NULL} (the default) will return the install folder for all
versions. If a version is specified (e.g \code{"2.11.4"}), it will return the path for a specific version.}
}
\value{
Path of install Pandoc binaries or specific version binary.
}
\description{
Path where Pandoc binaries are installed
}
