% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_is_active}
\alias{pandoc_is_active}
\title{Is a pandoc version active ?}
\usage{
pandoc_is_active(version)
}
\arguments{
\item{version}{This can be either:
\itemize{
\item \code{"latest"} for the latest release
\item A version number (e.g \code{"2.11.4"}) for a specific version
\item \code{"nightly"} for the last pandoc development built daily
}}
}
\description{
Is a pandoc version active ?
}
