% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-list.R
\name{pandoc_list_formats}
\alias{pandoc_list_formats}
\title{List available supported formats}
\usage{
pandoc_list_formats(type = c("input", "output"), bin = pandoc_bin())
}
\arguments{
\item{type}{Either list \code{input} or \code{output} formats. It corresponds to call
\code{--list-input-formats} and \code{--list-output-formats} respectively.}

\item{bin}{Path to a pandoc binary. Default to the active version.}
}
\value{
a data.frame (or a tibble if available) with 2 colums:
\itemize{
\item \code{type} (input or output)
\item \code{formats} (name of the formats that can be used as input or output)
}
}
\description{
List available supported formats
}
