% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_locate}
\alias{pandoc_locate}
\title{Locate a specific Pandoc version installed with this package}
\usage{
pandoc_locate(version = "default")
}
\arguments{
\item{version}{By default, it will return the path to the active pandoc
version. If a version is specified (e.g \code{"2.11.4"}), it will return the
path for a specific version if installed. If \code{latest}, the path the most
recent Pandoc version installed ('nightly' excluded). If \code{NULL}, the main
path to directory where all versions are stored.}
}
\value{
Path of Pandoc binaries root folder if installed.
}
\description{
This package helps install and manage Pandoc binaries in a specific folder.
This function helps with finding the path to thoses specific version of Pandoc.
See \code{\link[=pandoc_bin]{pandoc_bin()}} for another way of getting paths to \code{pandoc} binaries
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
