% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_locate}
\alias{pandoc_locate}
\title{Locate a specific Pandoc version installed with this package}
\usage{
pandoc_locate(version = "default")
}
\arguments{
\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}
}
\value{
Path of Pandoc binaries root folder if version is available.
}
\description{
This package helps install and manage Pandoc binaries in a specific folder.
This function helps with finding the path to thoses specific version of Pandoc.
See \code{\link[=pandoc_bin]{pandoc_bin()}} for another way of getting paths to \code{pandoc} binaries
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
