% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{pandoc_run}
\alias{pandoc_run}
\title{Run the pandoc binary from R}
\usage{
pandoc_run(args, bin = pandoc_bin(), echo = TRUE)
}
\arguments{
\item{args}{Character vector, arguments to the pandoc CLI command}

\item{bin}{Path to a pandoc binary. Default to the active version}

\item{echo}{Whether to rpint the standard output and error to the screen.}
}
\value{
The output of \code{\link[processx:run]{processx::run()}} invisibly
}
\description{
This function is a thin wrapper around the pandoc binary and allow to pass
any arguments supported by the Pandoc binary.
}
