% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{pandoc_run}
\alias{pandoc_run}
\title{Run the pandoc binary from R}
\usage{
pandoc_run(args, version = "default", echo = TRUE)
}
\arguments{
\item{args}{Character vector, arguments to the pandoc CLI command}

\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}

\item{echo}{Whether to print the standard output and error to the screen.}
}
\value{
The output of \code{\link[processx:run]{processx::run()}} invisibly
}
\description{
This function is a thin wrapper around the pandoc binary and allow to pass
any arguments supported by the Pandoc binary.
}
