% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_set_version}
\alias{pandoc_set_version}
\title{Activate a specific Pandoc version to be used}
\usage{
pandoc_set_version(version)
}
\value{
invisibly, the path to the previous active version
}
\description{
This function will set the specified version as the default version for the
session. By default, the default active version in the most recent one among
the installed version (nightly version excluded.)
}
