% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_set_version}
\alias{pandoc_set_version}
\title{Activate a specific Pandoc version to be used}
\usage{
pandoc_set_version(version)
}
\arguments{
\item{version}{This can be either:
\itemize{
\item \code{"latest"} for the latest release
\item A version number (e.g \code{"2.11.4"}) for a specific version
\item \code{"nightly"} for the last pandoc development built daily
}}
}
\value{
invisibly, the path to the previous active version
}
\description{
This function will set the specified version as the default version for the
session. By default, the default active version in the most recent one among
the installed version (nightly version excluded.)
}
