% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-available.R
\name{pandoc_set_version}
\alias{pandoc_set_version}
\title{Activate a specific Pandoc version to be used}
\usage{
pandoc_set_version(version, rmarkdown = TRUE)
}
\arguments{
\item{version}{This can be either:
\itemize{
\item \code{"latest"} for the latest release
\item A version number (e.g \code{"2.11.4"}) for a specific version
\item \code{"nightly"} for the last pandoc development built daily
}}

\item{rmarkdown}{if \code{TRUE} (the default) and \strong{rmarkdown} is available, this
will also set the pandoc version as the default one to use with
\strong{rmarkdown} by calling \code{\link[rmarkdown:find_pandoc]{rmarkdown::find_pandoc()}}}
}
\value{
invisibly, the previous active version.
}
\description{
This function will set the specified version as the default version for the
session. By default, the default active version in the most recent one among
the installed version (nightly version excluded.)
}
