% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_uninstall}
\alias{pandoc_uninstall}
\title{Uninstall a Pandoc version}
\usage{
pandoc_uninstall(version)
}
\arguments{
\item{version}{which version to uninstalled.}
}
\value{
\code{TRUE} (invisibly) if uninstallation is sucesfull.
}
\description{
You can run \code{\link[=pandoc_installed_versions]{pandoc_installed_versions()}} to see which versions are
currently installed on the system.
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
