% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_uninstall}
\alias{pandoc_uninstall}
\title{Uninstall a Pandoc version}
\usage{
pandoc_uninstall(version)
}
\arguments{
\item{version}{which version to uninstalled.}
}
\value{
\code{TRUE} (invisibly) if uninstalling is successful.
}
\description{
You can run \code{\link[=pandoc_installed_versions]{pandoc_installed_versions()}} to see which versions are
currently installed on the system.
}
\examples{
\dontshow{if (rlang::is_interactive() && !pandoc::pandoc_is_installed("2.19.2")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pandoc_install("2.19.2")
pandoc_is_installed("2.19.2")
pandoc_uninstall("2.19.2")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
