% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-install.R
\name{pandoc_update}
\alias{pandoc_update}
\title{Update to last Pandoc version available}
\usage{
pandoc_update()
}
\value{
invisibly, path to installed pandoc version
}
\description{
This function will check last version of Pandoc released, and install it if
not already installed.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pandoc_update()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=pandoc_install]{pandoc_install()}}
}
