% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rstudio_pandoc}
\alias{rstudio_pandoc}
\alias{pandoc_version_rstudio}
\alias{pandoc_bin_rstudio}
\title{Retrieve path and version of Pandoc shipped with RStudio}
\usage{
pandoc_version_rstudio()

pandoc_bin_rstudio()
}
\description{
RStudio IDE ships with a pandoc binary. The PATH is stored in \code{RSTUDIO_PANDOC}
environment variable. Theses function are helper to easily use this specific version.
}
