% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-bin.R
\name{rstudio_pandoc}
\alias{rstudio_pandoc}
\alias{pandoc_rstudio_version}
\alias{pandoc_rstudio_bin}
\title{Retrieve path and version of Pandoc shipped with RStudio}
\usage{
pandoc_rstudio_version()

pandoc_rstudio_bin()
}
\description{
RStudio IDE ships with a pandoc binary. The PATH is stored in \code{RSTUDIO_PANDOC}
environment variable. Theses function are helper to easily use this specific version.
}
