% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-available.R
\name{system_pandoc}
\alias{system_pandoc}
\alias{pandoc_system_version}
\alias{pandoc_system_bin}
\title{Retrieve path and version of Pandoc found on the system PATH}
\usage{
pandoc_system_version()

pandoc_system_bin()
}
\description{
Pandoc can also be installed on a system and available through the PATH.
Theses function are helper to easily use this specific version.
}
