% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{with_pandoc_version}
\alias{with_pandoc_version}
\title{Execute any code with a specific Pandoc version}
\usage{
with_pandoc_version(version, code, rmarkdown = FALSE)
}
\arguments{
\item{version}{Version to use. Default will be the \code{"default"} version. Other possible value are
\itemize{
\item A version number e.g \code{"2.14.1"}
\item The nightly version called \code{"nightly"}
\item The latest installed version with \code{"latest"}
\item Pandoc binary shipped with RStudio IDE with \code{"rstudio"}
\item Pandoc binary found in PATH with \code{"system"}
}}

\item{code}{Code to execute with the temporaty active Pandoc version.}

\item{rmarkdown}{if \code{TRUE} (the default) and \strong{rmarkdown} is available, this
will also set the pandoc version as the default one to use with
\strong{rmarkdown} by calling \code{\link[rmarkdown:find_pandoc]{rmarkdown::find_pandoc()}}}
}
\value{
The results of the evaluation of the \code{code} argument.
}
\description{
This function allows to run any R code by changing the active pandoc version to use
without modifying the R session state.
}
\details{
This is inspired from \strong{withr} package.
}
