% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-run.R
\name{with_pandoc_version}
\alias{with_pandoc_version}
\alias{local_pandoc_version}
\title{Execute any code with a specific Pandoc version}
\usage{
with_pandoc_version(version, code, rmarkdown = FALSE)

local_pandoc_version(version, rmarkdown = FALSE, .local_envir = parent.frame())
}
\arguments{
\item{version}{This can be either:
\itemize{
\item \code{"latest"} for the latest release
\item A version number (e.g \code{"2.11.4"}) for a specific version
\item \code{"nightly"} for the last pandoc development built daily
}}

\item{code}{Code to execute with the temporaty active Pandoc version.}

\item{rmarkdown}{if \code{TRUE} (the default) and \strong{rmarkdown} is available, this
will also set the pandoc version as the default one to use with
\strong{rmarkdown} by calling \code{\link[rmarkdown:find_pandoc]{rmarkdown::find_pandoc()}}}
}
\value{
The results of the evaluation of the \code{code} argument.
}
\description{
This function allows to run any R code by changing the active pandoc version to use
without modifying the R session state.
}
\details{
This is inspired from \strong{withr} package.
}
